requirements.txt
Django
django-environ
requests
pycountry